Combine BlockRegionTransform implementations into single class.
Reasoning: Allows for future serialization of transform attributes.
